+++ app/hooks/features/index.ts
@@ -717,36 +717,3 @@ export function useProjectFeatures(
     }
   );
 }
-
-export function useDownloadFeatureTemplate() {
-  const { data: session } = useSession();
-
-  const downloadFeatureTemplate = ({ pid }: { pid: Project['id'] }) => {
-    return DOWNLOADS.get<ArrayBuffer>(`/projects/${pid}/project-grid/shapefile-template`, {
-      responseType: 'arraybuffer',
-      headers: {
-        Authorization: `Bearer ${session.accessToken}`,
-        'Content-Type': 'application/zip',
-      },
-    });
-  };
-
-  return useMutation(downloadFeatureTemplate, {
-    onSuccess: (data, variables, context) => {
-      const { data: blob } = data;
-      const { pid } = variables;
-
-      const url = window.URL.createObjectURL(new Blob([blob]));
-      const link = document.createElement('a');
-      link.href = url;
-      link.setAttribute('download', `feature-template-${pid}.zip`);
-      document.body.appendChild(link);
-      link.click();
-      link.remove();
-      console.info('Success', data, variables, context);
-    },
-    onError: (error, variables, context) => {
-      console.info('Error', error, variables, context);
-    },
-  });
-}
