+++ app/hooks/projects/index.ts
@@ -15,6 +15,7 @@ import { Feature } from 'types/api/feature';
 import { Project } from 'types/api/project';
 import { createDownloadLink } from 'utils/download';
 
+import DOWNLOADS from 'services/downloads';
 import PROJECTS from 'services/projects';
 import UPLOADS from 'services/uploads';
 
@@ -788,5 +789,38 @@ export function useProjectTags(pid: Project['id']) {
       }).then((response) => response.data.data),
     enabled: !!pid,
     placeholderData: [],
+  });
+}
+
+export function useDownloadShapefileTemplate() {
+  const { data: session } = useSession();
+
+  const downloadShapefileTemplate = ({ pid }: { pid: Project['id'] }) => {
+    return DOWNLOADS.get<ArrayBuffer>(`/projects/${pid}/project-grid/shapefile-template`, {
+      responseType: 'arraybuffer',
+      headers: {
+        Authorization: `Bearer ${session.accessToken}`,
+        'Content-Type': 'application/zip',
+      },
+    });
+  };
+
+  return useMutation(downloadShapefileTemplate, {
+    onSuccess: (data, variables, context) => {
+      const { data: blob } = data;
+      const { pid } = variables;
+
+      const url = window.URL.createObjectURL(new Blob([blob]));
+      const link = document.createElement('a');
+      link.href = url;
+      link.setAttribute('download', `shapefile-template-${pid}.zip`);
+      document.body.appendChild(link);
+      link.click();
+      link.remove();
+      console.info('Success', data, variables, context);
+    },
+    onError: (error, variables, context) => {
+      console.info('Error', error, variables, context);
+    },
   });
 }
